requirements.txt
Flask==1.0.2
Flask-SocketIO==3.3.2
requests==2.21.0
speedtest-cli==2.1.1
eventlet==0.24.1
ruamel.yaml==0.15.89
